// src/generated/models/Erc_sla2sModel.ts
export const Erc_sla2serc_category = {
  1: 'Government',
  2: 'Insurance',
  3: 'Regulatory',
  4: 'Legal',
  5: 'Corporate',
  6: ' Partner'
} as const;
export type Erc_sla2serc_category = keyof typeof Erc_sla2serc_category;
export const Erc_sla2serc_daybasis = {
  1: 'Working Days',
  2: 'Calendar Days'
} as const;
export type Erc_sla2serc_daybasis = keyof typeof Erc_sla2serc_daybasis;
export const Erc_sla2serc_durationunit = {
  789180000: 'Working Days',
  789180001: 'Calendar Days',
  789180002: 'Hours',
  789180003: 'Minutes'
} as const;
export type Erc_sla2serc_durationunit = keyof typeof Erc_sla2serc_durationunit;
export const Erc_sla2serc_l1escalationtiming = {
  789180000: 'On Breach',
  789180001: 'Before Breach',
  789180002: 'After Breach'
} as const;
export type Erc_sla2serc_l1escalationtiming = keyof typeof Erc_sla2serc_l1escalationtiming;
export const Erc_sla2serc_l1recipientrole = {
  1: 'Supervisor',
  2: 'Dept Owner',
  3: 'Executive',
  4: 'BU Manager',
  5: 'Legal Lead'
} as const;
export type Erc_sla2serc_l1recipientrole = keyof typeof Erc_sla2serc_l1recipientrole;
export const Erc_sla2serc_l2escalationtiming = {
  789180000: 'On Breach',
  789180001: 'Before Breach',
  789180002: 'After Breach'
} as const;
export type Erc_sla2serc_l2escalationtiming = keyof typeof Erc_sla2serc_l2escalationtiming;
export const Erc_sla2serc_l2recipientrole = {
  789180000: ' Supervisor',
  789180001: 'Dept Owner',
  789180002: 'Executive',
  789180003: 'BU Manager',
  789180004: ' Legal Lead'
} as const;
export type Erc_sla2serc_l2recipientrole = keyof typeof Erc_sla2serc_l2recipientrole;
export const Erc_sla2serc_priority = {
  1: 'Critical',
  2: 'High',
  3: 'Medium',
  4: 'Low'
} as const;
export type Erc_sla2serc_priority = keyof typeof Erc_sla2serc_priority;
export const Erc_sla2serc_slastartpoint = {
  789180000: 'Received Date',
  789180001: 'Created Date',
  789180002: 'Assigned Date',
  789180003: 'Escalated Date'
} as const;
export type Erc_sla2serc_slastartpoint = keyof typeof Erc_sla2serc_slastartpoint;
export const Erc_sla2serc_slastoppoint = {
  789180000: 'Closed',
  789180001: 'Resolved',
  789180002: 'Cancelled',
  789180003: 'Expired'
} as const;
export type Erc_sla2serc_slastoppoint = keyof typeof Erc_sla2serc_slastoppoint;
export const Erc_sla2sstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Erc_sla2sstatecode = keyof typeof Erc_sla2sstatecode;
export const Erc_sla2sstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Erc_sla2sstatuscode = keyof typeof Erc_sla2sstatuscode;

export interface Erc_sla2sBase {
  erc_acknowledgementhours?: string;
  erc_active: boolean;
  "erc_AssignedOwner@odata.bind"?: string;
  erc_casereference?: string;
  erc_category?: Erc_sla2serc_category;
  erc_closedby?: string;
  erc_closeddate?: string;
  erc_communicationnumber?: string;
  erc_daybasis?: Erc_sla2serc_daybasis;
  erc_description?: string;
  erc_durationunit: Erc_sla2serc_durationunit;
  erc_evidenceattached?: boolean;
  erc_l1escalationtiming: Erc_sla2serc_l1escalationtiming;
  erc_l1recipientrole?: Erc_sla2serc_l1recipientrole;
  erc_l2escalationtiming?: Erc_sla2serc_l2escalationtiming;
  erc_l2recipientrole?: Erc_sla2serc_l2recipientrole;
  erc_name?: string;
  erc_priority?: Erc_sla2serc_priority;
  erc_reminderatpercent?: number;
  erc_responsedate?: string;
  erc_sla2id: string;
  erc_sladuration: number;
  erc_slastartpoint?: Erc_sla2serc_slastartpoint;
  erc_slastoppoint?: Erc_sla2serc_slastoppoint;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Erc_sla2sstatecode;
  statuscode?: Erc_sla2sstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Erc_sla2s extends Erc_sla2sBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  erc_activename?: string;
  erc_assignedownername?: string;
  erc_assignedowneryominame?: string;
  erc_categoryname?: string;
  erc_daybasisname?: string;
  erc_durationunitname?: string;
  erc_evidenceattachedname?: string;
  erc_l1escalationtimingname?: string;
  erc_l1recipientrolename?: string;
  erc_l2escalationtimingname?: string;
  erc_l2recipientrolename?: string;
  erc_priorityname?: string;
  erc_slastartpointname?: string;
  erc_slastoppointname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  erc_assignedowner?: object;
  _erc_assignedowner_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}