// src/generated/models/Erc_renewalsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Erc_renewalserc_escalationstatus = {
  1: 'Not Escalated',
  2: 'L1',
  3: 'L2'
} as const;
export type Erc_renewalserc_escalationstatus = keyof typeof Erc_renewalserc_escalationstatus;
export const Erc_renewalserc_renewalstatus = {
  1: 'Open',
  2: 'In progress',
  3: 'Completed',
  4: 'Overdue',
  5: 'Cancelled'
} as const;
export type Erc_renewalserc_renewalstatus = keyof typeof Erc_renewalserc_renewalstatus;
export const Erc_renewalserc_renewaltype = {
  1: 'License',
  2: 'Contract',
  3: 'Permit'
} as const;
export type Erc_renewalserc_renewaltype = keyof typeof Erc_renewalserc_renewaltype;
export const Erc_renewalserc_risklevel = {
  1: 'Low',
  2: 'Medium',
  3: 'High',
  4: 'Critical'
} as const;
export type Erc_renewalserc_risklevel = keyof typeof Erc_renewalserc_risklevel;
export const Erc_renewalsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Erc_renewalsstatecode = keyof typeof Erc_renewalsstatecode;
export const Erc_renewalsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Erc_renewalsstatuscode = keyof typeof Erc_renewalsstatuscode;

export interface Erc_renewalsBase {
  erc_completiondate?: string;
  erc_duedate?: string;
  erc_escalationstatus?: Erc_renewalserc_escalationstatus;
  erc_lastreminderdate?: string;
  "erc_Licenseandcontract@odata.bind"?: string;
  erc_newcolumn?: string;
  "erc_Owner@odata.bind"?: string;
  erc_remindercount?: number;
  erc_reminderdate?: string;
  erc_renewalid: string;
  erc_renewalnotes?: string;
  erc_renewalstatus?: Erc_renewalserc_renewalstatus;
  erc_renewaltype?: Erc_renewalserc_renewaltype;
  erc_risklevel?: Erc_renewalserc_risklevel;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  statecode: Erc_renewalsstatecode;
  statuscode?: Erc_renewalsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Erc_renewals extends Erc_renewalsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  erc_escalationstatusname?: string;
  erc_licenseandcontractname?: string;
  erc_ownername?: string;
  erc_owneryominame?: string;
  erc_renewalstatusname?: string;
  erc_renewaltypename?: string;
  erc_risklevelname?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  ownerid: string;
  owneridname: string;
  owneridtype: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  erc_licenseandcontract?: object;
  _erc_licenseandcontract_value?: string;
  erc_owner?: object;
  _erc_owner_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}